backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Lukz Financial Analysis Backend"
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "app"}]
[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.12.0"
slowapi = "^0.1.8"
fastapi = {extras = ["standard"], version = "^0.115.7"}
psycopg = {extras = ["binary"], version = "^3.2.4"}
httpx = "^0.28.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
